fix(conversations): move formatters to lib so tests don't pull streamdown
The previous commit made ConversationsTab import ChatMessage, which
transitively loads Streamdown/shiki (ESM). Jest couldn't parse them via
the component-under-test path. Move formatDuration and formatTimestamp
out of the component into lib/conversation-format.ts and import there
from both the component and the unit tests.
