+++ apps/web/src/components/chatbot/ConversationsTab.tsx
@@ -24,6 +24,10 @@ import { trpc, type RouterOutputs } from "@/lib/trpc";
 import { keepPreviousData } from "@tanstack/react-query";
 import { MessageSquare, Search, ArrowLeft, Clock } from "lucide-react";
 import { logError } from "@/lib/logger";
+import {
+  formatDuration,
+  formatTimestamp,
+} from "@/lib/conversation-format";
 
 type ConversationRow =
   RouterOutputs["analytics"]["getConversationsList"]["conversations"][number];
@@ -33,37 +37,6 @@ interface ConversationsTabProps {
 }
 
 type SortBy = "recent" | "mostMessages" | "longestDuration";
-
-function formatDuration(firstAt: Date | null, lastAt: Date | null): string {
-  if (!firstAt || !lastAt) return "-";
-  const ms = Math.max(
-    0,
-    new Date(lastAt).getTime() - new Date(firstAt).getTime(),
-  );
-  if (ms < 60000) return `${Math.round(ms / 1000)}s`;
-  if (ms < 3600000) return `${Math.round(ms / 60000)}m`;
-  const hours = Math.floor(ms / 3600000);
-  const mins = Math.round((ms % 3600000) / 60000);
-  return `${hours}h ${mins}m`;
-}
-
-function formatTimestamp(date: Date | string): string {
-  const d = new Date(date);
-  const now = new Date();
-  const diffMs = now.getTime() - d.getTime();
-  const diffDays = Math.floor(diffMs / 86400000);
-
-  if (diffDays === 0) {
-    return d.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
-  }
-  if (diffDays === 1) return "Yesterday";
-  if (diffDays < 7) return `${diffDays} days ago`;
-  return d.toLocaleDateString([], {
-    month: "short",
-    day: "numeric",
-    year: d.getFullYear() !== now.getFullYear() ? "numeric" : undefined,
-  });
-}
 
 function useDebouncedValue<T>(value: T, delayMs: number): T {
   const [debounced, setDebounced] = useState(value);
@@ -450,6 +423,3 @@ function ConversationDetail({
     </Card>
   );
 }
-
-// Exported pure helpers for unit tests.
-export { formatDuration, formatTimestamp };
